Added basic lighting and gui controls
The basic lighting lights the entire scene, all objects are no longer wireframes, the light within the sun lights every object except for the sun.

The three.min.js file has been added to allow GUI features however, the GUI in godmode is currently not working as time is a constant and changing it breaks the code.